corrections Sonar
