+++ app/components/home/dressingTab.component.tsx
@@ -1,0 +1,51 @@
+import DressingModel from "@/app/models/dressing.model";
+import { Image, Pressable, StyleSheet } from "react-native";
+import { ThemedView } from "../commons/ThemedView";
+import { ThemedText } from "../commons/ThemedText";
+import { Tabs } from "@/constants/TabsEnums";
+import { getTabIcon } from "../commons/tab/TabBarIcon";
+import { Colors } from "@/constants/Colors";
+
+
+
+/**
+ * @typedef DressingTabComponentProps
+ * @description Propriétés pour le composant de l'onglet Dressing.
+ * @property {DressingModel} dressing - Modèle représentant les données du dressing.
+ */
+type DressingTabComponentProps = {
+    dressing: DressingModel;
+    selectNewTab: (newTab: Tabs, _id?: string | undefined) => void;
+  };
+
+  const DressingTabComponent: React.FC<DressingTabComponentProps> = ({ dressing, selectNewTab }: DressingTabComponentProps) => {
+
+    return (
+      <Pressable onPress={() => selectNewTab(Tabs.DRESSING, dressing.id)}>
+        <ThemedView style={styles.container}>
+          <Image source={getTabIcon(true, dressing.categorie)} style={[styles.icon]} />
+          <ThemedText type="subtitle" style={{height: 40}}>{dressing.libelle}</ThemedText>
+        </ThemedView>
+      </Pressable>
+    );
+
+  }
+  export default DressingTabComponent;
+
+
+  const styles = StyleSheet.create({
+    container: {
+      zIndex: 0,
+      alignItems: 'center',
+      borderColor: 'grey',
+      borderWidth: 0.5
+    },
+  
+    icon: {
+      tintColor: 'gray',
+      margin: 5,
+      width: 250,
+      height: 250,
+      borderColor: Colors.dark.background,
+    }
+  });
